Fixed parsing of integers
string_view::data() must not be interpreted as a C-string (ie *not*
passed to strtol etc.), because it may not be terminated
properly. Example:

    string_view{"asd"}.substr(0,2).data()

is "asd", *not* "as".

This is an issue when fixed length literals are not separated in the
MML - for example
in (https://www.smwcentral.net/?p=section&a=details&id=18255),
"q7Ad16" was tokenized correctly as "q", "7A" - but then the
conversion to int returned 0x7ad16 instead of 0x7a.